+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Show abort status in tool execution
 - Improved dashboard generation speed and performance
 - Add to dashboard button to instantly add widgets to artifact
+- Added "Polish" dashboard for a quick update
 
 ## Version 0.0.353 (March 30, 2026)
 - feat: new a2a integration for timbr